# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MediFlow API
  version: "1.0.0"
  description: |
    Public surface of the MediFlow telemedicine platform. Every path below is
    served by the gateway on one port; behind it the requests fan out to five
    independent services, each with its own database.

    **Authentication.** Everything except `POST /auth/register` and `/health`
    needs a Keycloak-issued JWT. The gateway validates it at the edge and
    forwards it, and the service behind validates it again against the same
    JWKS, so a service reached directly is still protected.

    **Roles** live in the token's `realm_access.roles`: `patient`, `doctor`
    and `admin`. Where a role is required it is stated on the operation.

    **Reading someone else's appointment returns 404, not 403.** A 403 would
    confirm the appointment exists, which is itself a leak.
  license:
    name: MIT
    url: https://github.com/davidmm07/mediflow/blob/main/LICENSE
servers:
  - url: http://localhost:8080
    description: Local stack started with `make up`
...
security:
  - keycloak: []
  - bearerAuth: []
paths:
...
  /appointments:
    post:
      tags: [appointments]
      operationId: bookAppointment
      summary: Book an appointment
      description: |
        The one operation that spans an HTTP call and two databases. It
        reserves the slot in doctor-service before writing locally, because
        the reservation is the contended resource, and releases it again if
        the local write fails. On success it publishes `appointments.created`,
        which notification-service turns into a message.
      x-required-roles: [patient]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookRequest" }
      responses:
        "201":
          description: Appointment confirmed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: The practitioner or the slot does not exist
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Another patient took the slot first
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /appointments/me:
    get:
      tags: [appointments]
      operationId: listMyAppointments
      summary: List your own appointments
      responses:
        "200":
          description: Your appointments
          content:
            application/json:
              schema:
                type: object
                properties:
                  appointments:
                    type: array
                    items: { $ref: "#/components/schemas/Appointment" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /appointments/{appointmentID}:
    parameters:
      - $ref: "#/components/parameters/AppointmentID"
    get:
      tags: [appointments]
      operationId: getAppointment
      summary: Fetch one appointment
      description: |
        A caller who is neither the patient nor the practitioner gets **404,
        not 403**, because a 403 would confirm the appointment exists.
      responses:
        "200":
          description: The appointment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: No such appointment, or not yours
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /appointments/{appointmentID}/cancel:
    parameters:
      - $ref: "#/components/parameters/AppointmentID"
    post:
      tags: [appointments]
      operationId: cancelAppointment
      summary: Cancel an appointment
      description: |
        Open to the patient who booked it and to any `doctor` or `admin`. The
        local write happens before the slot is released, the opposite order to
        booking: a failed release costs a recoverable slot, while a failed
        local write would leave the patient believing a live appointment was
        cancelled.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CancelRequest" }
      responses:
        "200":
          description: The cancelled appointment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: No such appointment, or not yours
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: The appointment is not in a cancellable state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /doctors/{doctorID}/appointments:
    parameters:
      - $ref: "#/components/parameters/DoctorID"
    get:
      tags: [appointments]
      operationId: listDoctorAppointments
      summary: List a practitioner's appointments
      description: |
        An agenda is a sub-resource of the practitioner, so it is addressed
        under `/doctors`. appointment-service still owns the data; the gateway
        forwards this one nested path there rather than to doctor-service.
      x-required-roles: [doctor, admin]
      responses:
        "200":
          description: The practitioner's appointments
          content:
            application/json:
              schema:
                type: object
                properties:
                  appointments:
                    type: array
                    items: { $ref: "#/components/schemas/Appointment" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    keycloak:
      type: oauth2
      description: |
        The password grant is enabled on the `mediflow-gateway` client so this
        page and the smoke test can obtain a token unattended. A browser
        client should use Authorization Code with PKCE instead.
      flows:
        password:
          tokenUrl: http://localhost:8081/realms/mediflow/protocol/openid-connect/token
          scopes: {}
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Paste a token obtained from Keycloak directly.
  parameters:
    DoctorID:
      name: doctorID
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    AppointmentID:
      name: appointmentID
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: The body failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, malformed or expired token
      content:
        text/plain:
          schema: { type: string, example: invalid or expired token }
    Forbidden:
      description: The token is valid but lacks the required realm role
      content:
        text/plain:
          schema: { type: string, example: "forbidden: missing required role" }
...
  schemas:
    Error:
      type: object
      description: The uniform error body every service returns.
      properties:
        error: { type: string, example: doctor not found }
...
    BookRequest:
      type: object
      required: [doctor_id, slot_id]
      properties:
        doctor_id: { type: string, format: uuid }
        slot_id: { type: string, format: uuid }
        reason: { type: string, example: chest pain follow-up }
    CancelRequest:
      type: object
      properties:
        reason: { type: string, example: scheduling conflict }
    Appointment:
      type: object
      properties:
        id: { type: string, format: uuid }
        patient_user_id: { type: string, format: uuid }
        patient_name: { type: string }
        doctor_id: { type: string, format: uuid }
        doctor_name: { type: string }
        specialty: { type: string }
        slot_id: { type: string, format: uuid }
        starts_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time }
        status:
          type: string
          enum: [confirmed, cancelled, completed]
        reason: { type: string }
        cancelled_reason: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...